pyproject.toml
[project]
name = "demo"
version = "0.1.0"
description = ""
readme = "README.md"
authors = [
    { name = "Julian Hoever", email = "julianhoever@posteo.de" },
    { name = "Florian Hettstedt", email = "florian.hettstedt@uni-due.de"},
    { name = "David P. Federl", email = "david-peter.federl@uni-due.de"}
]
requires-python = ">=3.13"
dependencies = [
    "ipywidgets>=8.1.5",
    "jupyterlab>=4.3.5",
    "marimo>=0.15.0",
    "matplotlib>=3.10.0",
    "nicegui>=2.12.1",
    "numpy<2",
    "opencv-python>=4.11.0.86",
    "pandas>=2.2.3",
    "pillow>=11.1.0",
    "plotly>=6.3.0",
    "scikit-learn>=1.7.1",
    "torch>=2.6.0",
    "torchsummary>=1.5.1",
    "torchvision>=0.21.0",
]

[project.scripts]
demo = "demo.web_app:main"

[build-system]
requires = ["uv_build>=0.9.15,<0.10.0"]
build-backend = "uv_build"

[dependency-groups]
dev = [
    "mypy>=1.15.0",
    "pytest>=8.3.4",
    "python-lsp-server>=1.12.2",
    "rope>=1.13.0",
    "ruff>=0.9.7",
    "pynvim>=0.5.2",
    "ipykernel>=6.29.5",
]

[tool.marimo.runtime]
dotenv = [".env", ".env.testing"]
